Use double type for ncells

+++ examples/IBLevelSet/ex2/FlowGravityForcing.h
@@ -92,7 +92,7 @@ private:
     SAMRAI::tbox::Pointer<SAMRAI::pdat::CellVariable<NDIM, double> > d_ls_gas_var;
     double d_rho_neg, d_rho_pos;
     std::vector<double> d_grav_const;
-    int d_num_gas_interface_cells;
+    double d_num_gas_interface_cells;
 };
 
 //////////////////////////////////////////////////////////////////////////////
