+++ examples/wave_tank/ex1/example.cpp
@@ -289,8 +289,8 @@ main(int argc, char* argv[])
         const double rho_fluid = input_db->getDouble("RHO_F");
         const double rho_solid = input_db->getDoubleWithDefault("RHO_S", std::numeric_limits<double>::quiet_NaN());
         const double rho_gas = input_db->getDouble("RHO_G");
-        const int num_solid_interface_cells = input_db->getDouble("NUM_SOLID_INTERFACE_CELLS");
-        const int num_gas_interface_cells = input_db->getDouble("NUM_GAS_INTERFACE_CELLS");
+        const double num_solid_interface_cells = input_db->getDouble("NUM_SOLID_INTERFACE_CELLS");
+        const double num_gas_interface_cells = input_db->getDouble("NUM_GAS_INTERFACE_CELLS");
         const bool set_rho_solid = input_db->getBool("SET_RHO_S");
         SetFluidGasSolidDensity* ptr_setFluidGasSolidDensity = new SetFluidGasSolidDensity("SetFluidGasSolidDensity",
                                                                                            adv_diff_integrator,
